# Add option to build compiler with jemalloc