--- a/runtime/etc/Makefile.mem-jemalloc
+++ b/runtime/etc/Makefile.mem-jemalloc
@@ -16,4 +16,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
--include $(CHPL_MAKE_HOME)/runtime/etc/Makefile.jemalloc-$(CHPL_MAKE_JEMALLOC)
+-include $(CHPL_MAKE_HOME)/runtime/etc/Makefile.jemalloc-$(CHPL_MAKE_TARGET_JEMALLOC)