We are seeing a compile time reduction of around 18% for both hello world and larger compiles when using jemalloc instead of the system allocator.

* Adds `CHPL_HOST_{MEM,JEMALLOC}`
* Adds `CHPL_TARGET_{MEM,JEMALLOC}` which take their default from `CHPL_{MEM,JEMALLOC}` if not explicitly set. The former takes precedence over the latter

Platform differences as follows - more details in commit message:
* Linux: `CHPL_HOST_MEM=jemalloc` only supported when `CHPL_HOST_JEMALLOC=bundled`
* Darwin: `CHPL_HOST_MEM=jemalloc` only supported when `CHPL_HOST_JEMALLOC=system`
* Cygwin: jemalloc not supported (this was previously true)

Some example timings show a 10-25% reduction in compilation time.

wall time in seconds. best of 2

`chpl examples/hello.chpl`

|        | cstdlib | jemalloc |
|--------|--------:|---------:|
| clang  | 4.08    | 3.56     |
| llvm   | 3.96    | 3.41     |


`chpl test/regex/correctness.chpl`

|        | cstdlib | jemalloc |
|--------|--------:|---------:|
| clang  | 7.76    | 6.67     |
| llvm   | 7.40    | 6.26     |

`chpl --fast test/regex/correctness.chpl`

|        | cstdlib |jemalloc |
|--------|--------:|--------:|
| clang  | 17.51   | 16.60   |
| llvm   | 15.13   | 13.41   |

Arkouda on a Cray CS

|        | cstdlib | jemalloc |
|--------|--------:|---------:|
| gcc    | 563.64  | 503.90   |
| llvm   | 624.70  | 525.94   |

Tested on Linux and Mac