@ronawho 2021-11-03
There are almost certainly some doc changes required for this. I'd look for existing docs for `CHPL_MEM`/`CHPL_JEMALLOC`