@ronawho 2021-11-03
There's some logic that triggers a compiler rebuild if a component changed. Like if you switch from LLVM=none to bundled, we'll rebuild or something. I _think_ we may need to do that here too? I would guess Michael knows more about this.