@ronawho 2021-11-03
FYI here's the time to build on Arkouda on a Cray CS (16-node-cs-hdr) before and after this change:

before:
```
Pass               Name                   Main    Check    Clean     Time    %     Accum    %
---- ---------------------------------  -------  -------  -------  ------- -----  ------- -----
  42 makeBinary                         334.503    0.000    2.436  336.939  58.3  336.939  58.3
  13 resolve                            147.948    0.102    1.838  149.888  25.9  486.827  84.2
  22 parallel                            11.299    0.076    0.266   11.640   2.0  498.467  86.2
  20 callDestructors                     10.083    0.000    0.197   10.280   1.8  508.747  88.0
  29 copyPropagation                      8.645    0.057    0.794    9.497   1.6  518.244  89.6
  41 codegen                              7.710    0.000    0.171    7.881   1.4  526.125  91.0
  36 insertWideReferences                 7.423    0.051    0.206    7.680   1.3  533.805  92.3
  26 inlineFunctions                      5.743    0.122    0.967    6.832   1.2  540.637  93.5
  33 loopInvariantCodeMotion              5.791    0.048    0.121    5.960   1.0  546.596  94.6
  21 lowerIterators                       4.565    0.080    0.652    5.297   0.9  551.894  95.5
  ...
     total time                         563.640    1.264   13.174  578.078
```

after:
```
Pass               Name                   Main    Check    Clean     Time    %     Accum    %
---- ---------------------------------  -------  -------  -------  ------- -----  ------- -----
  42 makeBinary                         332.513    0.000    0.742  333.255  65.2  333.255  65.2
  13 resolve                            113.196    0.071    0.600  113.866  22.3  447.121  87.5
  22 parallel                             7.347    0.052    0.214    7.614   1.5  454.735  89.0
  20 callDestructors                      6.672    0.000    0.122    6.794   1.3  461.529  90.3
  36 insertWideReferences                 6.068    0.037    0.136    6.242   1.2  467.771  91.5
  29 copyPropagation                      5.772    0.042    0.308    6.122   1.2  473.892  92.7
  41 codegen                              5.567    0.000    0.125    5.692   1.1  479.584  93.8
  26 inlineFunctions                      4.705    0.078    0.562    5.346   1.0  484.930  94.9
  33 loopInvariantCodeMotion              4.531    0.033    0.102    4.667   0.9  489.596  95.8
  21 lowerIterators                       3.897    0.054    0.330    4.282   0.8  493.878  96.6
  ...
       total time                         503.901    0.897    6.236  511.035
```

This was with `CHPL_LLVM=none` if that matters. Also I'm wondering if there's any advantage to replacing the memory allocator for the C/llvm backend? For C backend we'd probably need a .so and `LD_LIBRARY_PATH`, but it's not clear to me if the current approach gets us using jemalloc with clang/llvm or if there'd be more to do there?