@aconsroe-hpe 2021-11-03
>  Also I'm wondering if there's any advantage to replacing the memory allocator for the C/llvm backend?

Yeah that's an interesting question. For gnu, the only way I can think of is `LD_PRELOAD` on Linux and even then, I'm not sure if there's only single process level (ie. can we be sure the env var gets propagated to all descendents?).

For builtin clang, we could build llvm+clang with jemalloc and try that out.