@aconsroe-hpe 2021-11-12
> There's some logic that triggers a compiler rebuild if a component changed. Like if you switch from LLVM=none to bundled, we'll rebuild or something. I think we may need to do that here too? I would guess Michael knows more about this.

This is now handled by [this commit](https://github.com/chapel-lang/chapel/pull/18627/commits/91f2f29a391c313ba8ee722f244df95f9cdd7f57)