@ronawho 2021-11-16
FYI here's timings for 16-node-cs-hdr with llvm this time instead of the c backend:

before:
```
Pass               Name                   Main    Check    Clean     Time    %     Accum    %
---- ---------------------------------  -------  -------  -------  ------- -----  ------- -----
  42 makeBinary                         350.803    0.000    2.477  353.280  55.2  353.280  55.2
  13 resolve                            157.137    0.110    1.897  159.144  24.9  512.425  80.1
  41 codegen                             39.173    0.000    0.134   39.307   6.1  551.732  86.3
  22 parallel                            12.525    0.080    0.268   12.874   2.0  564.606  88.3
  20 callDestructors                     11.554    0.000    0.220   11.775   1.8  576.380  90.1
  29 copyPropagation                      9.543    0.059    0.801   10.403   1.6  586.783  91.8
  36 insertWideReferences                 6.926    0.053    0.212    7.191   1.1  593.974  92.9
  26 inlineFunctions                      6.062    0.127    0.998    7.187   1.1  601.161  94.0
  33 loopInvariantCodeMotion              6.190    0.048    0.125    6.364   1.0  607.525  95.0
  21 lowerIterators                       4.765    0.083    0.676    5.525   0.9  613.050  95.9
  ...
     total time                         624.703    1.320   13.399  639.423
```

after:
```
Pass               Name                   Main    Check    Clean     Time    %     Accum    %
---- ---------------------------------  -------  -------  -------  ------- -----  ------- -----
  42 makeBinary                         303.885    0.000    0.806  304.691  57.1  304.691  57.1
  13 resolve                            131.109    0.073    0.635  131.816  24.7  436.507  81.8
  41 codegen                             33.943    0.000    0.114   34.057   6.4  470.564  88.2
  22 parallel                             8.701    0.053    0.220    8.973   1.7  479.537  89.9
  20 callDestructors                      6.808    0.000    0.125    6.933   1.3  486.470  91.2
  29 copyPropagation                      6.261    0.042    0.316    6.619   1.2  493.088  92.4
  36 insertWideReferences                 6.012    0.038    0.140    6.191   1.2  499.279  93.6
  33 loopInvariantCodeMotion              5.637    0.038    0.123    5.798   1.1  505.077  94.7
  26 inlineFunctions                      4.946    0.079    0.575    5.599   1.0  510.676  95.7
  21 lowerIterators                       4.041    0.055    0.337    4.432   0.8  515.109  96.6
  ...
     total time                         525.937    0.939    6.635  533.511
```

So more of an improvement since jemalloc is being used for parts of the backend too.