@ronawho 2021-11-16
Left a couple of inline comments, but overall I think this is in good shape. Can you add some perf results (maybe in table form) to the PR description to motivate the change?

I think the total time from 16-node-cs-hdr arkouda results and your numbers from https://github.com/Cray/chapel-private/issues/2558#issuecomment-970640581 would be great.