@aconsroe-hpe 2021-11-18
I've also added `test-host-jemalloc.bash` so we can add to nightly testing